requirements.txt
game==0.0.5
Pillow==8.4.0

playsound~=1.3.0
simpleaudio~=1.0.4